Use Travis CI to do automated testing on new code
Travis CI can be used to do automated testing on all pull requests before they are reviewed and merged. This service is free to open source projects like this one. The CI service must be turned on by @lzfelix going to https://travis-ci.com and logging in with GitHub credentials.